+++ backend/src/middleware/auth.ts
@@ -1,0 +1,28 @@
+import { Request, Response, NextFunction } from 'express';
+import type { AuthUser } from '../auth/types';
+
+export function requireAuth(req: Request, res: Response, next: NextFunction): void {
+  if (!req.user) {
+    res.status(401).json({ error: 'UNAUTHORIZED' });
+    return;
+  }
+  next();
+}
+
+export function requireManager(req: Request, res: Response, next: NextFunction): void {
+  const user = req.user as AuthUser | undefined;
+  if (!user || (user.role !== 'manager' && user.role !== 'admin')) {
+    res.status(403).json({ error: 'FORBIDDEN' });
+    return;
+  }
+  next();
+}
+
+export function requireAdmin(req: Request, res: Response, next: NextFunction): void {
+  const user = req.user as AuthUser | undefined;
+  if (!user || user.role !== 'admin') {
+    res.status(403).json({ error: 'FORBIDDEN' });
+    return;
+  }
+  next();
+}
